+++ src/common/net/Event.h
@@ -24,14 +24,14 @@ class Event {
      * @param timestamp in milliseconds since the beginning of the game, when
      * this action took place (in the point of view of the client)
      */
-    Event(quint32 id, quint64 timestamp);
+    Event(quint32 id, quint32 timestamp);
 
     /**
      * Creates a new Event with no ID (the ID will be set when the event is
      * sent)
      * @param timestamp in milliseconds since the beginning of the game.
      */
-    Event(quint64 timestamp);
+    Event(quint32 timestamp);
 
     virtual ~Event();
 
@@ -44,7 +44,7 @@ class Event {
 
   private:
     quint32 id_;
-    quint64 timestamp_;
+    quint32 timestamp_;
 
   protected:
     /**
